src/terminator/model/TextLine.java: switch to storing style information in a short[] instead of a byte[]...
src/terminator/model/StyledText.java: ...so we have room for two more bits,
controlling whether the foreground or background colors are set. This lets us
distinguish properly between a set color and a default color. In particular,
the colors in "whiptail --yesno food 8 40" are now correct, despite the fact
that our default color scheme is light-on-dark.

src/terminator/model/TextBuffer.java: 
src/terminator/terminal/TerminalListener.java: 
src/terminator/view/highlight/Highlight.java: track byte to short change.

src/terminator/terminal/escape/CSIEscapeAction.java: keep track of whether a
color has actually been set.

+++ terminator/trunk/src/terminator/view/highlight/Highlight.java
@@ -89,7 +89,7 @@ public class Highlight {
 			// style to be propagated to the RHS.  Also, if the highlight starts after the end
 			// of the visible text, we need to do the same.
 			if ((result.size() == 0) || highlightStartsAtEndOfLine(unlit, unlitStart)) {
-				StyledText empty = new StyledText("", (byte) StyledText.getDefaultStyle());
+				StyledText empty = new StyledText("", StyledText.getDefaultStyle());
 				result.add(new StyledText("", style.appliedTo(empty.getStyle())));
 			}
 			((StyledText) result.get(result.size() - 1)).setContinueToEnd(true);
